+++ libraries/Stopwatch_RT/StopWatch.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: StopWatch.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.3.6
+// VERSION: 0.4.0
 //    DATE: 2011-01-04
 // PURPOSE: Arduino Library implementing a stopwatch including seconds, milliseconds microseconds
 //     URL: https://github.com/RobTillaart/StopWatch_RT
@@ -49,6 +49,14 @@ void StopWatch::reset()
 }
 
 
+void StopWatch::restart()
+{
+  _state = StopWatch::RUNNING;
+  uint32_t t = _gettime();
+  _startTime = _stopTime = t;
+}
+
+
 //  made const as printTo expect a const object.
 uint32_t StopWatch::elapsed() const
 {
@@ -91,20 +99,20 @@ void StopWatch::setResolution(const enum Resolution resolution)
   switch(_resolution)
   {
   case MICROS:
-    _gettime = _micros;
+    _gettime = this->_micros;
     break;
 
   case SECONDS:
-    _gettime = seconds;
+    _gettime = this->seconds;
     break;
 
   case MINUTES:
-    _gettime = minutes;
+    _gettime = this->minutes;
     break;
 
   case MILLIS:
   default:
-    _gettime = _millis;
+    _gettime = this->_millis;
     _resolution = MILLIS;      //  for default!
     break;
   }